client: refacotr nav-toggle, added classnames removed header, added logo
