+++ package.json
@@ -33,6 +33,7 @@
   "dependencies": {
     "@babel/polyfill": "^7.0.0",
     "chalk": "^2.4.1",
+    "classnames": "^2.2.6",
     "compression": "^1.7.3",
     "connected-react-router": "^4.4.1",
     "cors": "^2.8.4",
@@ -73,7 +74,7 @@
     "copy-webpack-plugin": "^4.5.2",
     "css-loader": "^1.0.0",
     "eslint": "^5.5.0",
-    "eslint-config-google": "^0.9.1",
+    "eslint-config-google": "^0.10.0",
     "eslint-config-prettier": "^3.0.1",
     "eslint-formatter-pretty": "^1.3.0",
     "eslint-loader": "^2.1.0",
