--- src/client/components/Header/Header.module.scss
@@ -1,15 +1,0 @@
-.Header {
-  background: transparent;
-  height: 9em;
-
-  position: absolute;
-  left: 0;
-  right: 0;
-  top: 0;
-
-  display: flex;
-  justify-content: space-between;
-  align-items: center;
-
-  cursor: pointer;
-}
