--- src/client/components/Header/NavToggle.module.scss
@@ -1,65 +1,0 @@
-.NavToggle {
-  width: 26px;
-  height: 22px;
-  cursor: pointer;
-  margin-right: 2em;
-  padding: 25px;
-
-  &:hover .line {
-    &:first-child {
-      width: 22px;
-    }
-
-    &:nth-child(2) {
-      width: 22px;
-      margin-left: 4px;
-    }
-  }
-
-  .line {
-    width: 100%;
-    height: 3px;
-    display: block;
-    margin-bottom: 6px;
-    background: white;
-    border-radius: 2px;
-    position: relative;
-    top: 0;
-    transition: margin 0.3s ease, width 0.3s ease, background 0.3s ease, transform 0.3s ease 0.1s;
-
-    &:last-child {
-      margin: 0;
-    }
-  }
-}
-
-.navOpen {
-  width: 26px;
-  height: 22px;
-  cursor: pointer;
-  margin-right: 2em;
-  padding: 25px;
-  z-index: 100;
-
-  .line {
-    width: 27px;
-    height: 3px;
-    background-color: #000;
-    border-radius: 2px;
-    top: 50%;
-    position: absolute;
-    transition: margin 0.3s ease, background 0.3s ease, transform 0.3s ease 0.1s;
-
-    &:last-child {
-      transform: rotate(45deg);
-    }
-
-    &:nth-child(2) {
-      display: none;
-    }
-
-    &:first-child {
-      transform: rotate(-45deg);
-    }
-  }
-}
