+++ src/client/components/Logo/index.js
@@ -1,0 +1,11 @@
+import React from 'react';
+
+import styles from './Logo.module.scss';
+
+const Logo = () => {
+  return <div className={styles.Logo}>Logo</div>;
+};
+
+Logo.propTypes = {};
+
+export default Logo;
