+++ src/client/components/NavToggle/NavToggle.module.scss
@@ -1,0 +1,165 @@
+@import '../../styles/variables';
+
+.NavToggle {
+  position: absolute;
+  right: 20px;
+  top: 20px;
+  padding: 30px;
+  z-index: $nav-toggle-zindex;
+  outline: 0;
+  user-select: none;
+  transition: transform .2s ease;
+  cursor: pointer;
+
+  &:hover :not(.Open) {
+    .Bar:first-child {
+      width: 21px;
+    }
+
+    .Bar:nth-child(2) {
+      width: 23px;
+    }
+  }
+}
+
+@media (min-width: $lg) {
+  .NavToggle {
+    right:40px;
+    top: 40px
+  }
+}
+
+.Bar {
+  width: 27px;
+  height: 3px;
+  background-color: #fff;
+  border-radius: 2px;
+  margin: auto;
+  top: 50%;
+  right: 0;
+  position: absolute;
+  transition: margin .1s, width .1s, background .1s .1s;
+  transition-timing-function: ease-out;
+
+  &:first-child {
+    right: auto;
+    left: 0;
+    transition: margin .1s,width .3s,background .1s .1s;
+    margin-top: -9px;
+    animation: runHamburgerTopBar .75s .15s;
+    animation-timing-function: ease
+  }
+
+  &:nth-child(3) {
+    margin-top: 9px;
+    animation: runHamburgerBottomBar .75s .15s;
+    animation-timing-function: ease
+  }
+}
+
+.BarWrapper {
+  width: 27px;
+  height: 100%;
+  position: absolute;
+  top: 0;
+  left: 50%;
+  transform: translateX(-50%) translateY(0);
+}
+
+.Open {
+  .Bar {
+    background-color: #000;
+    transition: margin .15s, width .15s, background .1s .1s;
+
+    &:first-child {
+      width: 27px;
+      margin-top: 0;
+      animation: crossHamburgerTopBar .3s .15s forwards;
+      animation-timing-function: ease-out;
+    }
+
+    &:nth-child(2) {
+      width: 0
+    }
+
+    &:nth-child(3) {
+      margin-top: 0;
+      animation: crossHamburgerBottomBar .3s .15s forwards;
+      animation-timing-function: ease-out
+    }
+  }
+}
+
+
+@keyframes runHamburgerTopBar {
+  0% {
+      transform: rotate(-4deg)
+  }
+
+  25% {
+      transform: rotate(4deg)
+  }
+
+  50% {
+      transform: rotate(-2deg)
+  }
+
+  to {
+      transform: rotate(0deg)
+  }
+}
+
+
+@keyframes runHamburgerBottomBar {
+  0% {
+      transform: rotate(4deg)
+  }
+
+  25% {
+      transform: rotate(-4deg)
+  }
+
+  50% {
+      transform: rotate(2deg)
+  }
+
+  to {
+      transform: rotate(0deg)
+  }
+}
+
+@keyframes crossHamburgerTopBar {
+  0% {
+      transform: rotate(0deg)
+  }
+
+  25% {
+      transform: rotate(-55deg)
+  }
+
+  50% {
+      transform: rotate(-40deg)
+  }
+
+  to {
+      transform: rotate(-45deg)
+  }
+}
+
+@keyframes crossHamburgerBottomBar {
+  0% {
+      transform: rotate(0deg)
+  }
+
+  25% {
+      transform: rotate(50deg)
+  }
+
+  50% {
+      transform: rotate(40deg)
+  }
+
+  to {
+      transform: rotate(45deg)
+  }
+}
