+++ src/client/index.html
@@ -1,15 +1,15 @@
 <!DOCTYPE html>
 <html lang="en">
 
-<head>
-  <meta charset="utf-8">
-  <title>oneDeeds</title>
-  <meta name="viewport" content="width=device-width, initial-scale=1">
-  <base href="/" />
-</head>
+  <head>
+    <meta charset="utf-8">
+    <title>oneDeeds</title>
+    <meta name="viewport" content="width=device-width, initial-scale=1">
+    <base href="/" />
+  </head>
 
-<body>
-  <div id="root">Loading...</div>
-</body>
+  <body>
+    <div id="root">Loading...</div>
+  </body>
 
 </html>
