+++ src/client/pages/Home/Home.module.scss
@@ -1,1 +1,5 @@
 @import '../../styles/variables';
+
+.Home {
+  height: 100vh;
+}
