+++ src/client/pages/Home/index.js
@@ -3,14 +3,16 @@ import { connect } from 'react-redux';
 
 import styles from './Home.module.scss';
 
-import Header from 'Components/Header';
+import NavToggle from 'Components/NavToggle';
 import Layout from 'Components/Layout';
 import Nav from 'Components/Nav';
+import Logo from 'Components/Logo';
 
 const Home = () => {
   return (
     <div className={styles.Home}>
-      <Header />
+      <NavToggle />
+      <Logo />
       <Nav />
       <Layout />
     </div>
