+++ src/client/styles/global.scss
@@ -1,2 +1,28 @@
+// App
 @import './variables';
 @import './base';
+
+
+// // Example
+// // Extra small devices (portrait phones, less than 576px)
+// // No media query since this is the default in Bootstrap
+
+// // Small devices (landscape phones, 576px and up)
+// @media (min-width: $sm) {
+
+// }
+
+// // Medium devices (tablets, 768px and up)
+// @media (min-width: $md) {
+
+// }
+
+// // Large devices (desktops, 992px and up)
+// @media (min-width: $lg) {
+
+// }
+
+// // Extra large devices (large desktops, 1200px and up)
+// @media (min-width: $xl) {
+
+// }
